sampling time considers the simulation time

+++ src/luposdate3000_simulator_iot/src/jvmMain/kotlin/Device.kt
@@ -23,7 +23,7 @@ public class Device(
     public var processedSensorDataPackages: Long = 0
         private set
 
-    private fun getNetworkDelay(destinationAddress: Int): Int {
+    private fun getNetworkDelay(destinationAddress: Int): Long {
         return if (destinationAddress == address) {
             0
         } else {
